refactor: isolate hook-dependent simple fees tests
