+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip1261/CryptoTransferSimpleFeesTest.java
@@ -17,14 +17,11 @@ import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAliasedAccountInfo;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
 import static com.hedera.services.bdd.spec.queries.crypto.ExpectedTokenRel.relationshipWith;
-import static com.hedera.services.bdd.spec.transactions.TxnUtils.accountAllowanceHook;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.mintToken;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenAssociate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFee;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.moving;
 import static com.hedera.services.bdd.spec.transactions.token.TokenMovement.movingHbar;
@@ -110,9 +107,6 @@ public class CryptoTransferSimpleFeesTest {
     private static final String PAYER = "payer";
     private static final String THRESHOLD_PAYER = "thresholdPayer";
     private static final String PAYER_INSUFFICIENT_BALANCE = "payerInsufficientBalance";
-    private static final String PAYER_WITH_HOOK = "payerWithHook";
-    private static final String PAYER_WITH_TWO_HOOKS = "payerWithTwoHooks";
-    private static final String PAYER_WITH_THREE_HOOKS = "payerWithThreeHooks";
     private static final String RECEIVER_ASSOCIATED_FIRST = "receiverAssociatedFirst";
     private static final String RECEIVER_ASSOCIATED_SECOND = "receiverAssociatedSecond";
     private static final String RECEIVER_ASSOCIATED_THIRD = "receiverAssociatedThird";
@@ -137,13 +131,9 @@ public class CryptoTransferSimpleFeesTest {
     private static final String adminKey = "adminKey";
     private static final String supplyKey = "supplyKey";
     private static final String PAYER_KEY = "payerKey";
-    private static final String HOOK_CONTRACT = "TruePreHook";
-
     @BeforeAll
     static void beforeAll(@NonNull final TestLifecycle testLifecycle) {
-        testLifecycle.overrideInClass(Map.of(
-                "fees.simpleFeesEnabled", "true",
-                "hooks.hooksEnabled", "true"));
+        testLifecycle.overrideInClass(Map.of("fees.simpleFeesEnabled", "true"));
     }
 
     @Nested
@@ -2154,332 +2144,6 @@ public class CryptoTransferSimpleFeesTest {
                                     .hasTokenBalance(FUNGIBLE_TOKEN_2, 90L)));
                 }
             }
-
-            @Nested
-            @DisplayName("Crypto Transfer With Hooks - Simple Fees Positive Tests")
-            class CryptoTransferWithHooksSimpleFeesPositiveTests {
-                @HapiTest
-                @DisplayName("Crypto Transfer HBAR with hook execution - extra hook full charging")
-                final Stream<DynamicTest> cryptoTransferHBARWithOneHookExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-
-                            // transfer tokens
-                            cryptoTransfer(movingHbar(1L).between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK)
-                                    .fee(ONE_MILLION_HBARS)
-                                    .via("hbarTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "hbarTransferTxn",
-                                    txnSize -> expectedCryptoTransferHbarFullFeeUsd(Map.of(
-                                            SIGNATURES, 1L,
-                                            HOOK_EXECUTION, 1L,
-                                            ACCOUNTS, 2L,
-                                            GAS, 5_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001)));
-                }
-
-                @HapiTest
-                @DisplayName("Crypto Transfer FT with hook execution - extra hook full charging")
-                final Stream<DynamicTest> cryptoTransferFTWithOneHookExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_HOOK, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_FIRST, FUNGIBLE_TOKEN),
-
-                            // transfer tokens
-                            cryptoTransfer(moving(10L, FUNGIBLE_TOKEN)
-                                            .between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("ftTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "ftTransferTxn",
-                                    txnSize -> expectedCryptoTransferFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 1L,
-                                            HOOK_EXECUTION, 1L,
-                                            ACCOUNTS, 2L,
-                                            TOKEN_TYPES, 1L,
-                                            GAS, 5_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_HOOK).hasTokenBalance(FUNGIBLE_TOKEN, 90L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_FIRST).hasTokenBalance(FUNGIBLE_TOKEN, 10L)));
-                }
-
-                @HapiTest
-                @DisplayName("Crypto Transfer FT with same hook executed twice - extra hook full charging")
-                final Stream<DynamicTest> cryptoTransferFTWithOneHookExecutedTwiceExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_HOOK, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_FIRST, FUNGIBLE_TOKEN),
-                            tokenAssociate(RECEIVER_ASSOCIATED_SECOND, FUNGIBLE_TOKEN),
-
-                            // transfer tokens
-                            cryptoTransfer(
-                                            moving(10L, FUNGIBLE_TOKEN)
-                                                    .between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST),
-                                            moving(10L, FUNGIBLE_TOKEN)
-                                                    .between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_SECOND))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("ftTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "ftTransferTxn",
-                                    txnSize -> expectedCryptoTransferFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 1L,
-                                            HOOK_EXECUTION, 1L,
-                                            ACCOUNTS, 3L,
-                                            TOKEN_TYPES, 1L,
-                                            GAS, 5_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_HOOK).hasTokenBalance(FUNGIBLE_TOKEN, 80L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_FIRST).hasTokenBalance(FUNGIBLE_TOKEN, 10L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_SECOND).hasTokenBalance(FUNGIBLE_TOKEN, 10L)));
-                }
-
-                @HapiTest
-                @DisplayName("Crypto Transfer NFT with hook execution - extra hook full charging")
-                final Stream<DynamicTest> cryptoTransferNFTWithOneHookExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createNonFungibleTokenWithoutCustomFees(
-                                    NON_FUNGIBLE_TOKEN, PAYER_WITH_HOOK, supplyKey, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_FIRST, NON_FUNGIBLE_TOKEN),
-                            mintNFT(NON_FUNGIBLE_TOKEN, 1, 5),
-
-                            // transfer tokens
-                            cryptoTransfer(movingUnique(NON_FUNGIBLE_TOKEN, 1L)
-                                            .between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST))
-                                    .withNftSenderPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("nftTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "nftTransferTxn",
-                                    txnSize -> expectedCryptoTransferNFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 1L,
-                                            HOOK_EXECUTION, 1L,
-                                            ACCOUNTS, 2L,
-                                            TOKEN_TYPES, 1L,
-                                            GAS, 5_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_HOOK).hasTokenBalance(NON_FUNGIBLE_TOKEN, 3L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_FIRST).hasTokenBalance(NON_FUNGIBLE_TOKEN, 1L)));
-                }
-
-                @HapiTest
-                @DisplayName("Crypto Transfer HBAR and FT with hook execution - extra hooks full charging")
-                final Stream<DynamicTest> cryptoTransferHBARAndFtWithTwoHooksExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_TWO_HOOKS, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_FIRST, FUNGIBLE_TOKEN),
-
-                            // transfer tokens
-                            cryptoTransfer(
-                                            movingHbar(1L).between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST),
-                                            moving(10L, FUNGIBLE_TOKEN)
-                                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .withPreHookFor(PAYER_WITH_TWO_HOOKS, 2L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK, PAYER_WITH_TWO_HOOKS)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("tokenTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "tokenTransferTxn",
-                                    txnSize -> expectedCryptoTransferHBARAndFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 2L,
-                                            HOOK_EXECUTION, 2L,
-                                            ACCOUNTS, 3L,
-                                            TOKEN_TYPES, 1L,
-                                            GAS, 10_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_TWO_HOOKS).hasTokenBalance(FUNGIBLE_TOKEN, 90L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_FIRST).hasTokenBalance(FUNGIBLE_TOKEN, 10L)));
-                }
-
-                @HapiTest
-                @DisplayName("Crypto Transfer HBAR and NFT with hook execution - extra hooks full charging")
-                final Stream<DynamicTest> cryptoTransferHBARAndNFtWithOneHookExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createNonFungibleTokenWithoutCustomFees(
-                                    NON_FUNGIBLE_TOKEN, PAYER_WITH_TWO_HOOKS, supplyKey, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_FIRST, NON_FUNGIBLE_TOKEN),
-                            mintNFT(NON_FUNGIBLE_TOKEN, 1, 5),
-
-                            // transfer tokens
-                            cryptoTransfer(
-                                            movingHbar(1L).between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST),
-                                            movingUnique(NON_FUNGIBLE_TOKEN, 1L)
-                                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 2L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK, PAYER_WITH_TWO_HOOKS)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("tokenTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "tokenTransferTxn",
-                                    txnSize -> expectedCryptoTransferHBARAndNFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 2L,
-                                            HOOK_EXECUTION, 2L,
-                                            ACCOUNTS, 3L,
-                                            TOKEN_TYPES, 1L,
-                                            GAS, 10_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_TWO_HOOKS).hasTokenBalance(NON_FUNGIBLE_TOKEN, 3L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_FIRST).hasTokenBalance(NON_FUNGIBLE_TOKEN, 1L)));
-                }
-
-                @HapiTest
-                @DisplayName(
-                        "Crypto Transfer HBAR, FT and NFT with hook execution - extra hooks and accounts full charging")
-                final Stream<DynamicTest> cryptoTransferHBARAndFtAndNFTWithTwoHooksExtraHooksAndAccountsFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_TWO_HOOKS, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_SECOND, FUNGIBLE_TOKEN),
-                            createNonFungibleTokenWithoutCustomFees(
-                                    NON_FUNGIBLE_TOKEN, PAYER_WITH_TWO_HOOKS, supplyKey, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_THIRD, NON_FUNGIBLE_TOKEN),
-                            mintNFT(NON_FUNGIBLE_TOKEN, 1, 5),
-
-                            // transfer tokens
-                            cryptoTransfer(
-                                            movingHbar(1L).between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST),
-                                            moving(10L, FUNGIBLE_TOKEN)
-                                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_SECOND),
-                                            movingUnique(NON_FUNGIBLE_TOKEN, 1L)
-                                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_THIRD))
-                                    .withPreHookFor(PAYER_WITH_TWO_HOOKS, 1L, 5_000_000L, "")
-                                    .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 2L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK, PAYER_WITH_TWO_HOOKS)
-                                    .fee(ONE_MILLION_HBARS)
-                                    .via("tokenTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "tokenTransferTxn",
-                                    txnSize -> expectedCryptoTransferHBARAndFTAndNFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 2L,
-                                            HOOK_EXECUTION, 2L,
-                                            ACCOUNTS, 5L,
-                                            TOKEN_TYPES, 2L,
-                                            GAS, 10_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_TWO_HOOKS)
-                                    .hasTokenBalance(FUNGIBLE_TOKEN, 90L)
-                                    .hasTokenBalance(NON_FUNGIBLE_TOKEN, 3L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_SECOND).hasTokenBalance(FUNGIBLE_TOKEN, 10L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_THIRD).hasTokenBalance(NON_FUNGIBLE_TOKEN, 1L)));
-                }
-
-                @HapiTest
-                @DisplayName(
-                        "Crypto Transfer HBAR, FT and NFT with hook execution - extra hooks, tokens and accounts full charging")
-                final Stream<DynamicTest>
-                        cryptoTransferHBARAndFtAndNFTWithTwoHooksExtraHooksAndTokensAndAccountsFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_TWO_HOOKS, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_SECOND, FUNGIBLE_TOKEN),
-                            createNonFungibleTokenWithoutCustomFees(
-                                    NON_FUNGIBLE_TOKEN, PAYER_WITH_TWO_HOOKS, supplyKey, adminKey),
-                            tokenAssociate(RECEIVER_ASSOCIATED_THIRD, NON_FUNGIBLE_TOKEN),
-                            mintNFT(NON_FUNGIBLE_TOKEN, 1, 5),
-
-                            // transfer tokens
-                            cryptoTransfer(
-                                            movingHbar(1L).between(PAYER_WITH_HOOK, RECEIVER_ASSOCIATED_FIRST),
-                                            moving(10L, FUNGIBLE_TOKEN)
-                                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_SECOND),
-                                            movingUnique(NON_FUNGIBLE_TOKEN, 1L)
-                                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_THIRD))
-                                    .withPreHookFor(PAYER_WITH_TWO_HOOKS, 1L, 5_000_000L, "")
-                                    .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 2L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK, PAYER_WITH_TWO_HOOKS)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("tokenTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "tokenTransferTxn",
-                                    txnSize -> expectedCryptoTransferHBARAndFTAndNFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 2L,
-                                            HOOK_EXECUTION, 2L,
-                                            ACCOUNTS, 5L,
-                                            TOKEN_TYPES, 2L,
-                                            GAS, 10_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            getAccountBalance(PAYER_WITH_TWO_HOOKS)
-                                    .hasTokenBalance(FUNGIBLE_TOKEN, 90L)
-                                    .hasTokenBalance(NON_FUNGIBLE_TOKEN, 3L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_SECOND).hasTokenBalance(FUNGIBLE_TOKEN, 10L),
-                            getAccountBalance(RECEIVER_ASSOCIATED_THIRD).hasTokenBalance(NON_FUNGIBLE_TOKEN, 1L)));
-                }
-
-                @HapiTest
-                @DisplayName(
-                        "Crypto Transfer - Auto Create Accounts with FT moving and with hook execution - extra hook full charging")
-                final Stream<DynamicTest> cryptoTransferFTAutoAccountCreationWithOneHookExtraHookFullCharging() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_HOOK, adminKey),
-
-                            // transfer tokens
-                            cryptoTransfer(moving(10L, FUNGIBLE_TOKEN).between(PAYER_WITH_HOOK, VALID_ALIAS_ED25519))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 5_000_000L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("ftTransferTxn"),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "ftTransferTxn",
-                                    txnSize -> expectedCryptoTransferFTFullFeeUsd(Map.of(
-                                                    SIGNATURES, 1L,
-                                                    HOOK_EXECUTION, 1L,
-                                                    ACCOUNTS, 2L,
-                                                    TOKEN_TYPES, 1L,
-                                                    GAS, 5_000_000L,
-                                                    PROCESSING_BYTES, (long) txnSize))
-                                            + TOKEN_ASSOCIATE_EXTRA_FEE_USD,
-                                    0.001),
-                            // validate auto-created account properties
-                            getAliasedAccountInfo(VALID_ALIAS_ED25519)
-                                    .hasToken(relationshipWith(FUNGIBLE_TOKEN))
-                                    .has(accountWith()
-                                            .key(VALID_ALIAS_ED25519)
-                                            .alias(VALID_ALIAS_ED25519)
-                                            .maxAutoAssociations(-1)),
-                            // validate balances
-                            getAccountBalance(PAYER_WITH_HOOK).hasTokenBalance(FUNGIBLE_TOKEN, 90L)));
-                }
-            }
         }
 
         @Nested
@@ -3801,38 +3465,6 @@ public class CryptoTransferSimpleFeesTest {
                                         checkOwnerBalance);
                             })));
                 }
-
-                @HapiTest
-                @DisplayName("Crypto Transfer - Auto Create Accounts with FT moving and failing hook - fails on handle")
-                final Stream<DynamicTest> cryptoTransferFTAutoAccountCreationWithFailingHookFailsOnHandle() {
-                    return hapiTest(flattened(
-                            // create keys, tokens and accounts
-                            createAccountsAndKeys(),
-                            createFungibleTokenWithoutCustomFees(FUNGIBLE_TOKEN, 100L, PAYER_WITH_HOOK, adminKey),
-
-                            // transfer tokens
-                            cryptoTransfer(moving(10L, FUNGIBLE_TOKEN).between(PAYER_WITH_HOOK, VALID_ALIAS_ED25519))
-                                    .withPreHookFor(PAYER_WITH_HOOK, 1L, 10L, "")
-                                    .payingWith(PAYER_WITH_HOOK)
-                                    .signedBy(PAYER_WITH_HOOK)
-                                    .fee(ONE_HUNDRED_HBARS)
-                                    .via("ftTransferTxn")
-                                    .hasKnownStatus(INSUFFICIENT_GAS),
-                            validateChargedUsdWithinWithTxnSize(
-                                    "ftTransferTxn",
-                                    txnSize -> expectedCryptoTransferFTFullFeeUsd(Map.of(
-                                            SIGNATURES, 1L,
-                                            HOOK_EXECUTION, 1L,
-                                            ACCOUNTS, 2L,
-                                            TOKEN_TYPES, 1L,
-                                            GAS, 10L,
-                                            PROCESSING_BYTES, (long) txnSize)),
-                                    0.001),
-                            // validate no auto-created account exists
-                            getAliasedAccountInfo(VALID_ALIAS_ED25519).hasCostAnswerPrecheck(INVALID_ACCOUNT_ID),
-                            // validate balances
-                            getAccountBalance(PAYER_WITH_HOOK).hasTokenBalance(FUNGIBLE_TOKEN, 100L)));
-                }
             }
         }
 
@@ -3880,20 +3512,6 @@ public class CryptoTransferSimpleFeesTest {
             return List.of(
                     cryptoCreate(PAYER).balance(ONE_HUNDRED_HBARS),
                     cryptoCreate(PAYER_INSUFFICIENT_BALANCE).balance(ONE_HBAR / 100000),
-                    uploadInitCode(HOOK_CONTRACT),
-                    contractCreate(HOOK_CONTRACT).gas(5_000_000),
-                    cryptoCreate(PAYER_WITH_HOOK)
-                            .balance(ONE_MILLION_HBARS)
-                            .withHook(accountAllowanceHook(1L, HOOK_CONTRACT)),
-                    cryptoCreate(PAYER_WITH_TWO_HOOKS)
-                            .balance(ONE_HUNDRED_HBARS)
-                            .withHook(accountAllowanceHook(1L, HOOK_CONTRACT))
-                            .withHook(accountAllowanceHook(2L, HOOK_CONTRACT)),
-                    cryptoCreate(PAYER_WITH_THREE_HOOKS)
-                            .balance(ONE_HUNDRED_HBARS)
-                            .withHook(accountAllowanceHook(1L, HOOK_CONTRACT))
-                            .withHook(accountAllowanceHook(2L, HOOK_CONTRACT))
-                            .withHook(accountAllowanceHook(3L, HOOK_CONTRACT)),
                     cryptoCreate(OWNER).balance(ONE_HUNDRED_HBARS),
                     cryptoCreate(HBAR_OWNER_INSUFFICIENT_BALANCE).balance(ONE_HBAR / 100000),
                     cryptoCreate(RECEIVER_ASSOCIATED_FIRST).balance(ONE_HBAR),
