+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip1261/CryptoTransferWithCustomFeesSimpleFeesTest.java
@@ -4,14 +4,11 @@ package com.hedera.services.bdd.suites.hip1261;
 import static com.hedera.services.bdd.junit.TestTags.SIMPLE_FEES;
 import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getAccountBalance;
-import static com.hedera.services.bdd.spec.transactions.TxnUtils.accountAllowanceHook;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoTransfer;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.mintToken;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenAssociate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
-import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFee;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHbarFeeInheritingRoyaltyCollector;
 import static com.hedera.services.bdd.spec.transactions.token.CustomFeeSpecs.fixedHtsFee;
@@ -40,7 +37,6 @@ import static org.hiero.hapi.support.fees.Extra.HOOK_EXECUTION;
 import static org.hiero.hapi.support.fees.Extra.PROCESSING_BYTES;
 import static org.hiero.hapi.support.fees.Extra.SIGNATURES;
 import static org.hiero.hapi.support.fees.Extra.TOKEN_TYPES;
-
 import com.google.protobuf.ByteString;
 import com.hedera.services.bdd.junit.HapiTest;
 import com.hedera.services.bdd.junit.HapiTestLifecycle;
@@ -99,17 +95,11 @@ public class CryptoTransferWithCustomFeesSimpleFeesTest {
     private static final String feeScheduleKey = "feeScheduleKey";
     private static final String supplyKey = "supplyKey";
 
-    private static final String HOOK_CONTRACT = "TruePreHook";
-    private static final String PAYER_WITH_HOOK = "payerWithHook";
-    private static final String PAYER_WITH_TWO_HOOKS = "payerWithTwoHooks";
-
     private static final String DUMMY_TOKEN = "dummyToken";
 
     @BeforeAll
     static void beforeAll(@NonNull final TestLifecycle testLifecycle) {
-        testLifecycle.overrideInClass(Map.of(
-                "fees.simpleFeesEnabled", "true",
-                "hooks.hooksEnabled", "true"));
+        testLifecycle.overrideInClass(Map.of("fees.simpleFeesEnabled", "true"));
     }
 
     @Nested
@@ -1011,92 +1001,6 @@ public class CryptoTransferWithCustomFeesSimpleFeesTest {
                     getAccountBalance(RECEIVER_ASSOCIATED_SECOND).hasTokenBalance(FT_WITH_HBAR_FIXED_FEE, 10L),
                     getAccountBalance(RECEIVER_FREE_AUTO_ASSOCIATIONS).hasTokenBalance(FT_WITH_HBAR_FIXED_FEE, 5L)));
         }
-
-        @HapiTest
-        @DisplayName(
-                "Crypto Transfer FT and NFT with custom fees and with hook execution - extra hooks, tokens and accounts full charging")
-        final Stream<DynamicTest>
-                cryptoTransferHBARAndFtAndNFTWithTwoHooksExtraHooksAndTokensAndAccountsFullCharging() {
-            return hapiTest(flattened(
-                    // create keys, tokens and accounts
-                    createAccountsAndKeys(),
-                    createFungibleTokenWithAdminKey(DENOM_TOKEN, 100, DENOM_TREASURY, adminKey),
-                    tokenAssociate(DENOM_COLLECTOR, DENOM_TOKEN),
-                    tokenAssociate(PAYER_WITH_TWO_HOOKS, DENOM_TOKEN),
-                    tokenAssociate(RECEIVER_ASSOCIATED_FIRST, DENOM_TOKEN),
-                    tokenAssociate(RECEIVER_ASSOCIATED_THIRD, DENOM_TOKEN),
-                    cryptoTransfer(
-                            moving(10, DENOM_TOKEN).between(DENOM_TREASURY, PAYER_WITH_TWO_HOOKS),
-                            moving(10, DENOM_TOKEN).between(DENOM_TREASURY, RECEIVER_ASSOCIATED_FIRST),
-                            moving(10, DENOM_TOKEN).between(DENOM_TREASURY, RECEIVER_ASSOCIATED_THIRD)),
-                    createFungibleTokenWithFixedHtsFee(
-                            FT_WITH_HTS_FIXED_FEE, 100L, PAYER_WITH_TWO_HOOKS, adminKey, HTS_FEE),
-                    tokenAssociate(HTS_COLLECTOR, FT_WITH_HTS_FIXED_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_FIRST, FT_WITH_HTS_FIXED_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_SECOND, FT_WITH_HTS_FIXED_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_THIRD, FT_WITH_HTS_FIXED_FEE),
-                    cryptoTransfer(
-                            moving(10L, FT_WITH_HTS_FIXED_FEE).between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST),
-                            moving(10L, FT_WITH_HTS_FIXED_FEE)
-                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_THIRD)),
-                    createNFTWithFixedFee(NFT_WITH_HTS_FEE, PAYER_WITH_TWO_HOOKS, supplyKey, adminKey, HTS_FEE),
-                    mintNFT(NFT_WITH_HTS_FEE, 0, 5),
-                    tokenAssociate(RECEIVER_ASSOCIATED_FIRST, NFT_WITH_HTS_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_SECOND, NFT_WITH_HTS_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_THIRD, NFT_WITH_HTS_FEE),
-
-                    // transfer tokens
-                    cryptoTransfer(
-                            moving(20L, FT_WITH_HTS_FIXED_FEE).between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST),
-                            movingUnique(NFT_WITH_HTS_FEE, 1L, 2L, 3L)
-                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_THIRD)),
-                    cryptoTransfer(
-                                    moving(10L, FT_WITH_HTS_FIXED_FEE)
-                                            .between(RECEIVER_ASSOCIATED_FIRST, RECEIVER_ASSOCIATED_SECOND),
-                                    movingUnique(NFT_WITH_HTS_FEE, 1L)
-                                            .between(RECEIVER_ASSOCIATED_THIRD, RECEIVER_ASSOCIATED_FIRST),
-                                    movingUnique(NFT_WITH_HTS_FEE, 2L)
-                                            .between(RECEIVER_ASSOCIATED_THIRD, RECEIVER_ASSOCIATED_SECOND),
-                                    movingUnique(NFT_WITH_HTS_FEE, 4L)
-                                            .between(PAYER_WITH_TWO_HOOKS, RECEIVER_FREE_AUTO_ASSOCIATIONS),
-                                    movingUnique(NFT_WITH_HTS_FEE, 5L)
-                                            .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST))
-                            .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 1L, 5_000_000L, "")
-                            .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 2L, 5_000_000L, "")
-                            .payingWith(PAYER_WITH_TWO_HOOKS)
-                            .signedBy(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST, RECEIVER_ASSOCIATED_THIRD)
-                            .fee(ONE_HUNDRED_HBARS)
-                            .via("tokenTransferTxn"),
-                    validateChargedUsdWithinWithTxnSize(
-                            "tokenTransferTxn",
-                            txnSize -> expectedCryptoTransferTokenWithCustomFullFeeUsd(Map.of(
-                                            SIGNATURES, 3L,
-                                            HOOK_EXECUTION, 2L,
-                                            ACCOUNTS, 5L,
-                                            TOKEN_TYPES, 5L,
-                                            GAS, 10_000_000L,
-                                            PROCESSING_BYTES, (long) txnSize))
-                                    + TOKEN_ASSOCIATE_EXTRA_FEE_USD,
-                            0.001),
-                    getAccountBalance(HTS_COLLECTOR).hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 2L),
-                    getAccountBalance(DENOM_COLLECTOR).hasTokenBalance(DENOM_TOKEN, 2L),
-                    getAccountBalance(PAYER_WITH_TWO_HOOKS)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 60L)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 0L)
-                            .hasTokenBalance(DENOM_TOKEN, 10L),
-                    getAccountBalance(RECEIVER_ASSOCIATED_FIRST)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 20L)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 2L)
-                            .hasTokenBalance(DENOM_TOKEN, 9L),
-                    getAccountBalance(RECEIVER_ASSOCIATED_SECOND)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 10L)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 1L),
-                    getAccountBalance(RECEIVER_ASSOCIATED_THIRD)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 1L)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 8L)
-                            .hasTokenBalance(DENOM_TOKEN, 9L),
-                    getAccountBalance(RECEIVER_FREE_AUTO_ASSOCIATIONS).hasTokenBalance(NFT_WITH_HTS_FEE, 1L)));
-        }
     }
 
     @Nested
@@ -1287,90 +1191,6 @@ public class CryptoTransferWithCustomFeesSimpleFeesTest {
                     getAccountBalance(RECEIVER_ASSOCIATED_SECOND).hasTokenBalance(FT_WITH_HBAR_FIXED_FEE, 0L),
                     getAccountBalance(RECEIVER_FREE_AUTO_ASSOCIATIONS).hasTokenBalance(FT_WITH_HBAR_FIXED_FEE, 0L)));
         }
-
-        @HapiTest
-        @DisplayName("Crypto Transfer FT and NFT with custom fees and with failing hook - fails on handle")
-        final Stream<DynamicTest> cryptoTransferHBARAndFtAndNFTWithCustomFeesAndWithFailingHookFailsOnHandle() {
-            return hapiTest(flattened(
-                    // create keys, tokens and accounts
-                    createAccountsAndKeys(),
-                    createFungibleTokenWithAdminKey(DENOM_TOKEN, 100, DENOM_TREASURY, adminKey),
-                    tokenAssociate(DENOM_COLLECTOR, DENOM_TOKEN),
-                    tokenAssociate(PAYER_WITH_TWO_HOOKS, DENOM_TOKEN),
-                    tokenAssociate(RECEIVER_ASSOCIATED_FIRST, DENOM_TOKEN),
-                    tokenAssociate(RECEIVER_ASSOCIATED_THIRD, DENOM_TOKEN),
-                    cryptoTransfer(
-                            moving(10, DENOM_TOKEN).between(DENOM_TREASURY, PAYER_WITH_TWO_HOOKS),
-                            moving(10, DENOM_TOKEN).between(DENOM_TREASURY, RECEIVER_ASSOCIATED_FIRST),
-                            moving(10, DENOM_TOKEN).between(DENOM_TREASURY, RECEIVER_ASSOCIATED_THIRD)),
-                    createFungibleTokenWithFixedHtsFee(
-                            FT_WITH_HTS_FIXED_FEE, 100L, PAYER_WITH_TWO_HOOKS, adminKey, HTS_FEE),
-                    tokenAssociate(HTS_COLLECTOR, FT_WITH_HTS_FIXED_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_FIRST, FT_WITH_HTS_FIXED_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_SECOND, FT_WITH_HTS_FIXED_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_THIRD, FT_WITH_HTS_FIXED_FEE),
-                    cryptoTransfer(
-                            moving(10L, FT_WITH_HTS_FIXED_FEE).between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST),
-                            moving(10L, FT_WITH_HTS_FIXED_FEE)
-                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_THIRD)),
-                    createNFTWithFixedFee(NFT_WITH_HTS_FEE, PAYER_WITH_TWO_HOOKS, supplyKey, adminKey, HTS_FEE),
-                    mintNFT(NFT_WITH_HTS_FEE, 0, 5),
-                    tokenAssociate(RECEIVER_ASSOCIATED_FIRST, NFT_WITH_HTS_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_SECOND, NFT_WITH_HTS_FEE),
-                    tokenAssociate(RECEIVER_ASSOCIATED_THIRD, NFT_WITH_HTS_FEE),
-
-                    // transfer tokens
-                    cryptoTransfer(
-                            moving(20L, FT_WITH_HTS_FIXED_FEE).between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST),
-                            movingUnique(NFT_WITH_HTS_FEE, 1L, 2L, 3L)
-                                    .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_THIRD)),
-                    cryptoTransfer(
-                                    moving(10L, FT_WITH_HTS_FIXED_FEE)
-                                            .between(RECEIVER_ASSOCIATED_FIRST, RECEIVER_ASSOCIATED_SECOND),
-                                    movingUnique(NFT_WITH_HTS_FEE, 1L)
-                                            .between(RECEIVER_ASSOCIATED_THIRD, RECEIVER_ASSOCIATED_FIRST),
-                                    movingUnique(NFT_WITH_HTS_FEE, 2L)
-                                            .between(RECEIVER_ASSOCIATED_THIRD, RECEIVER_ASSOCIATED_SECOND),
-                                    movingUnique(NFT_WITH_HTS_FEE, 4L)
-                                            .between(PAYER_WITH_TWO_HOOKS, RECEIVER_FREE_AUTO_ASSOCIATIONS),
-                                    movingUnique(NFT_WITH_HTS_FEE, 5L)
-                                            .between(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST))
-                            .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 1L, 10L, "")
-                            .withNftSenderPreHookFor(PAYER_WITH_TWO_HOOKS, 2L, 10L, "")
-                            .payingWith(PAYER_WITH_TWO_HOOKS)
-                            .signedBy(PAYER_WITH_TWO_HOOKS, RECEIVER_ASSOCIATED_FIRST, RECEIVER_ASSOCIATED_THIRD)
-                            .fee(ONE_HUNDRED_HBARS)
-                            .via("tokenTransferTxn")
-                            .hasKnownStatus(INSUFFICIENT_GAS),
-                    validateChargedUsdWithinWithTxnSize(
-                            "tokenTransferTxn",
-                            txnSize -> expectedCryptoTransferTokenWithCustomFullFeeUsd(Map.of(
-                                    SIGNATURES, 3L,
-                                    HOOK_EXECUTION, 2L,
-                                    ACCOUNTS, 5L,
-                                    TOKEN_TYPES, 5L,
-                                    GAS, 20L,
-                                    PROCESSING_BYTES, (long) txnSize)),
-                            0.001),
-                    getAccountBalance(HTS_COLLECTOR).hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 0L),
-                    getAccountBalance(DENOM_COLLECTOR).hasTokenBalance(DENOM_TOKEN, 0L),
-                    getAccountBalance(PAYER_WITH_TWO_HOOKS)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 60L)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 2L)
-                            .hasTokenBalance(DENOM_TOKEN, 10L),
-                    getAccountBalance(RECEIVER_ASSOCIATED_FIRST)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 30L)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 0L)
-                            .hasTokenBalance(DENOM_TOKEN, 10L),
-                    getAccountBalance(RECEIVER_ASSOCIATED_SECOND)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 0L)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 0L),
-                    getAccountBalance(RECEIVER_ASSOCIATED_THIRD)
-                            .hasTokenBalance(NFT_WITH_HTS_FEE, 3L)
-                            .hasTokenBalance(FT_WITH_HTS_FIXED_FEE, 10L)
-                            .hasTokenBalance(DENOM_TOKEN, 10L),
-                    getAccountBalance(RECEIVER_FREE_AUTO_ASSOCIATIONS).hasTokenBalance(NFT_WITH_HTS_FEE, 0L)));
-        }
     }
 
     private HapiTokenCreate createFungibleTokenWithFixedHbarFee(
@@ -1489,15 +1309,6 @@ public class CryptoTransferWithCustomFeesSimpleFeesTest {
         return List.of(
                 cryptoCreate(PAYER).balance(ONE_HUNDRED_HBARS),
                 cryptoCreate(PAYER_INSUFFICIENT_BALANCE).balance(ONE_HBAR / 100000),
-                uploadInitCode(HOOK_CONTRACT),
-                contractCreate(HOOK_CONTRACT).gas(5_000_000),
-                cryptoCreate(PAYER_WITH_HOOK)
-                        .balance(ONE_HUNDRED_HBARS)
-                        .withHook(accountAllowanceHook(1L, HOOK_CONTRACT)),
-                cryptoCreate(PAYER_WITH_TWO_HOOKS)
-                        .balance(ONE_HUNDRED_HBARS)
-                        .withHook(accountAllowanceHook(1L, HOOK_CONTRACT))
-                        .withHook(accountAllowanceHook(2L, HOOK_CONTRACT)),
                 cryptoCreate(OWNER).balance(ONE_MILLION_HBARS),
                 cryptoCreate(NEW_TREASURY_WITH_UNLIMITED_AUTO_ASSOCIATIONS)
                         .balance(ONE_HUNDRED_HBARS)
